+++ src/testFixtures/java/org/opensearch/neuralsearch/OpenSearchSecureRestTestCase.java
@@ -2,25 +2,39 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.opensearch.neuralsearch;
+
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManager;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManagerBuilder;
+import org.apache.hc.client5.http.ssl.ClientTlsStrategyBuilder;
+import org.apache.hc.core5.http.nio.ssl.TlsStrategy;
+import org.apache.hc.core5.util.Timeout;
+import static org.opensearch.client.RestClientBuilder.DEFAULT_MAX_CONN_PER_ROUTE;
+import static org.opensearch.client.RestClientBuilder.DEFAULT_MAX_CONN_TOTAL;
+import static org.opensearch.knn.common.KNNConstants.MODEL_INDEX_NAME;
+import static org.opensearch.neuralsearch.TestUtils.NEURAL_SEARCH_BWC_PREFIX;
+import static org.opensearch.neuralsearch.TestUtils.OPENDISTRO_SECURITY;
+import static org.opensearch.neuralsearch.TestUtils.OPENSEARCH_SYSTEM_INDEX_PREFIX;
+import static org.opensearch.neuralsearch.TestUtils.SECURITY_AUDITLOG_PREFIX;
+import static org.opensearch.neuralsearch.TestUtils.SKIP_DELETE_MODEL_INDEX;
 
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.apache.http.Header;
-import org.apache.http.HttpHost;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.message.BasicHeader;
-import org.apache.http.ssl.SSLContextBuilder;
+import org.apache.hc.client5.http.auth.AuthScope;
+import org.apache.hc.client5.http.auth.UsernamePasswordCredentials;
+import org.apache.hc.client5.http.impl.auth.BasicCredentialsProvider;
+import org.apache.hc.client5.http.ssl.NoopHostnameVerifier;
+import org.apache.hc.core5.http.Header;
+import org.apache.hc.core5.http.HttpHost;
+import org.apache.hc.core5.http.ParseException;
+import org.apache.hc.core5.http.message.BasicHeader;
+import org.apache.hc.core5.ssl.SSLContextBuilder;
 import org.junit.After;
 import org.opensearch.client.Request;
 import org.opensearch.client.Response;
@@ -50,6 +64,12 @@ public abstract class OpenSearchSecureRestTestCase extends OpenSearchRestTestCas
     private static final String DEFAULT_SOCKET_TIMEOUT = "60s";
     private static final String INTERNAL_INDICES_PREFIX = ".";
     private static String protocol;
+
+    private final Set<String> IMMUTABLE_INDEX_PREFIXES = Set.of(
+        NEURAL_SEARCH_BWC_PREFIX,
+        SECURITY_AUDITLOG_PREFIX,
+        OPENSEARCH_SYSTEM_INDEX_PREFIX
+    );
 
     @Override
     protected String getProtocol() {
@@ -97,13 +117,20 @@ public abstract class OpenSearchSecureRestTestCase extends OpenSearchRestTestCas
                 .orElseThrow(() -> new RuntimeException("user name is missing"));
             final String password = Optional.ofNullable(System.getProperty(SYS_PROPERTY_KEY_PASSWORD))
                 .orElseThrow(() -> new RuntimeException("password is missing"));
-            final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-            credentialsProvider.setCredentials(AuthScope.ANY, new UsernamePasswordCredentials(userName, password));
+            final BasicCredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+            final AuthScope anyScope = new AuthScope(null, -1);
+            credentialsProvider.setCredentials(anyScope, new UsernamePasswordCredentials(userName, password.toCharArray()));
             try {
-                return httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider)
-                    // disable the certificate since our testing cluster just uses the default security configuration
-                    .setSSLHostnameVerifier(NoopHostnameVerifier.INSTANCE)
-                    .setSSLContext(SSLContextBuilder.create().loadTrustMaterial(null, (chains, authType) -> true).build());
+                final TlsStrategy tlsStrategy = ClientTlsStrategyBuilder.create()
+                    .setHostnameVerifier(NoopHostnameVerifier.INSTANCE)
+                    .setSslContext(SSLContextBuilder.create().loadTrustMaterial(null, (chains, authType) -> true).build())
+                    .build();
+                final PoolingAsyncClientConnectionManager connectionManager = PoolingAsyncClientConnectionManagerBuilder.create()
+                    .setMaxConnPerRoute(DEFAULT_MAX_CONN_PER_ROUTE)
+                    .setMaxConnTotal(DEFAULT_MAX_CONN_TOTAL)
+                    .setTlsStrategy(tlsStrategy)
+                    .build();
+                return httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider).setConnectionManager(connectionManager);
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
@@ -114,7 +141,12 @@ public abstract class OpenSearchSecureRestTestCase extends OpenSearchRestTestCas
             socketTimeoutString == null ? DEFAULT_SOCKET_TIMEOUT : socketTimeoutString,
             CLIENT_SOCKET_TIMEOUT
         );
-        builder.setRequestConfigCallback(conf -> conf.setSocketTimeout(Math.toIntExact(socketTimeout.getMillis())));
+        builder.setRequestConfigCallback(conf -> {
+            Timeout timeout = Timeout.ofMilliseconds(Math.toIntExact(socketTimeout.getMillis()));
+            conf.setConnectTimeout(timeout);
+            conf.setResponseTimeout(timeout);
+            return conf;
+        });
         if (settings.hasValue(CLIENT_PATH_PREFIX)) {
             builder.setPathPrefix(settings.get(CLIENT_PATH_PREFIX));
         }
@@ -129,7 +161,7 @@ public abstract class OpenSearchSecureRestTestCase extends OpenSearchRestTestCas
     }
 
     @After
-    public void deleteExternalIndices() throws IOException {
+    public void deleteExternalIndices() throws IOException, ParseException {
         final Response response = client().performRequest(new Request("GET", "/_cat/indices?format=json" + "&expand_wildcards=all"));
         try (
             final XContentParser parser = JsonXContent.jsonXContent.createParser(
@@ -153,8 +185,29 @@ public abstract class OpenSearchSecureRestTestCase extends OpenSearchRestTestCas
                 .collect(Collectors.toList());
 
             for (final String indexName : externalIndices) {
-                adminClient().performRequest(new Request("DELETE", "/" + indexName));
+                if (!skipDeleteIndex(indexName)) {
+                    adminClient().performRequest(new Request("DELETE", "/" + indexName));
+                }
             }
         }
+    }
+
+    private boolean getSkipDeleteModelIndexFlag() {
+        return Boolean.parseBoolean(System.getProperty(SKIP_DELETE_MODEL_INDEX, "false"));
+    }
+
+    private boolean skipDeleteModelIndex(String indexName) {
+        return (MODEL_INDEX_NAME.equals(indexName) && getSkipDeleteModelIndexFlag());
+    }
+
+    private boolean skipDeleteIndex(String indexName) {
+        if (indexName != null
+            && !OPENDISTRO_SECURITY.equals(indexName)
+            && IMMUTABLE_INDEX_PREFIXES.stream().noneMatch(indexName::startsWith)
+            && !skipDeleteModelIndex(indexName)) {
+            return false;
+        }
+
+        return true;
     }
 }